Fix some checks and mismatched types

+++ include/megaapi_impl.h
@@ -331,7 +331,7 @@ class MegaRequestPrivate : public MegaRequest
 		void setFile(const char* file);
         void setParamType(int type);
         void setText(const char* text);
-        void setNumber(int number);
+        void setNumber(long long number);
         void setFlag(bool flag);
         void setTransferTag(int transfer);
         void setListener(MegaRequestListener *listener);
